network: automatically retry binding the listener socket
It is possible that binding to a socket will fail at the program
startup, for example because DHCP has not fetched the IP address for the
interface, or the interface is down (because WiFi).

This changes the `Network` class to implement retries every 5 seconds if
the initial bind has failed. Based on my experiments once the first bind
succeeds, the socket will remain properly bound and working even if the
interface loses the IP address or disappears entirely, so that is nice,
but unknown how reliable this behaviour is. In particular my intuition
was screming to me that `accept` call would raise an error which we
would use to set `self.__bind_socket = None`, so that we'd know we need
to rebind later.

While here I had to make some changes to the structure, for example:

* I had to replace `self.__running: bool` with `self.__stop:
  threading.Event`, as that was required to have an interruptible sleep
  in the `__get_bound_socket` function.
* I replaced the "self-connect" with socket shutdown in order to advance
  `accept` loop, as there isn't a guarantee that the network was
  listening on `127.0.0.1` in the first place!
* I also made the SSL context creation part of the very same retry logic
  – now if somebody configures invalid SSL certificates, or if those
  certificates are made available at a later date (e.g. fetched from
  letsencrypt?) the velbustcp server won’t get stuck in a similar
  situation as it was getting stuck with socket binds.